requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
et-xmlfile==1.1.0
fastapi==0.115.0
greenlet==3.0.3
h11==0.14.0
idna==3.10
MouseInfo==0.1.3
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
playwright==1.47.0
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PyGetWindow==0.0.9
pymongo==4.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
ruff==0.7.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
